Refactored the most important alien methods in game to AlienController
